+++ R/simulateFactorModelNullsFromSingularValuesAndLoadings.R
@@ -11,15 +11,14 @@
 ##' vector of \code{k} factors}
 ##' @author Gilles Blanchard, Pierre Neuvial and Etienne Roquain
 ##' @importFrom Matrix Matrix
-##' @export
 ##' @examples
 ##'
 ##' m <- 10
 ##'
 ##' ## independent
-##' sim0 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m)
+##' sim0 <- sansSouci:::simulateFactorModelNullsFromSingularValuesAndLoadings(m)
 ##' str(sim0)
-##' sum(is.na(simulateFactorModelNullsFromSingularValuesAndLoadings(m)))
+##' sum(is.na(sansSouci:::simulateFactorModelNullsFromSingularValuesAndLoadings(m)))
 ##' S0 <- getFactorModelCovarianceMatrix(m)
 ##' image(S0)
 ##'
@@ -27,7 +26,7 @@
 ##' rho <- 0.2
 ##' h <- 1
 ##' P <- matrix(1, m, length(h))
-##' sim1 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
+##' sim1 <- sansSouci:::simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
 ##' str(sim1)
 ##' S1 <- getFactorModelCovarianceMatrix(m, h, P, rho=rho)
 ##' image(S1)
@@ -41,7 +40,7 @@
 ##' gamma3 <- rep(c(-1, 1), times=m/2)
 ##' P <- cbind(gamma1, gamma2, gamma3)/sqrt(m)
 ##'
-##' sim3 <- simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
+##' sim3 <- sansSouci:::simulateFactorModelNullsFromSingularValuesAndLoadings(m, h, P, rho=rho)
 ##' str(sim3)
 ##' S3 <- getFactorModelCovarianceMatrix(m, h, P, rho=rho)
 ##' image(S3)